Changed channel to not use async send for messages to clients.
The reason is that the current client protocol (and the one they likely want us to use?) is to always respond to each request. If we were to use async, the sending process will get a response for each message back but will never actuallty read them. This would leave a lot of messages in its "mailbox".

Might not be a problem if mailbox is discarded on process exit, which it probably is... Should look up.